requirements.txt
flask
flasgger
flask_cors
python-dotenv
mysql-connector-python
marshmallow